% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{create_palette}
\usage{
create_palette(path, n_colors)
}
\arguments{
\item{path}{The path to the image to be used for the palette.}

\item{n_colors}{The size of the palette.}
}
\value{
A vector of hex codes.
}
\description{
This is a function to pull colors from an image.
}
