% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_plvs_vltra.R
\name{display_plvs_vltra}
\alias{display_plvs_vltra}
\title{display_plvs_vltra}
\usage{
display_plvs_vltra(palette_name)
}
\arguments{
\item{palette_name}{The palette to be selected.}
}
\value{
A plot object.
}
\description{
This is a function to display a pre-made palette.
}
