% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plvs_vltra.R
\name{plvs_vltra}
\alias{plvs_vltra}
\title{plvs_vltra}
\usage{
plvs_vltra(palette_name, show = FALSE)
}
\arguments{
\item{palette_name}{The palette to be selected.}

\item{show}{Set to TRUE to plot the pallete using "scales".}
}
\value{
A vector of hex codes.
}
\description{
This is a function to select a pre-made palette.
}
