% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plvs_vltra_palettes.R
\name{plvs_vltra_palettes}
\alias{plvs_vltra_palettes}
\title{plvs_vltra_palettes}
\usage{
plvs_vltra_palettes(palette_name, list_all = FALSE)
}
\arguments{
\item{palette_name}{The palette to be selected.}

\item{list_all}{Set to TRUE to return all palette names.}
}
\value{
A vector of hex codes.
}
\description{
This is a function to select a pre-made palette.
}
