% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plvsvltra.R
\docType{package}
\name{PlvsVltra}
\alias{PlvsVltra}
\title{PlvsVltra: Beautiful scientific data visualization palletes}
\description{
The package, PlvsVltra, provides beautiful scientific data visualization palletes based on
the color combinations found in Spanish and Moroccan tiles. There are two key functions: create_palette()
can be used to create palletes of up to eight colors using a point-and-click interface to select
hex codes from an image; plvs_vltra_palletes() can be used to load one of the pre-made palletes 
displayed on the readme page.
}
